+++ core-services/service-request/src/main/java/org/egov/servicerequest/web/models/AttributeDefinition.java
@@ -62,6 +62,8 @@ public class AttributeDefinition {
 
         MULTIVALUELIST("MultiValueList"),
 
+        NUMERIC("Numeric"),
+
         FILE("File");
 
         private String value;