+++ core-services/service-request/CHANGELOG.md
@@ -1,5 +1,11 @@
 All notable changes to this module will be documented in this file.
 
+## 1.1.0 - 2025-06-11
+- Added update service request api support
+- Added `clientReferenceId` and `serviceClientReferenceId` for `AttributeValue`
+- Introduced `Numeric` data type for `AttributeDefinition`
+- Validation for service request for client id already exists on create
+
 ## 1.0.1 - 2024-08-29 
 
 - Added `BOOLEAN` DataType in `AttributeDefinition`